+++ src/builder/mac.ts
@@ -1,0 +1,10 @@
+import {Complete} from './complete';
+
+export interface Mac {
+
+    /**
+     * Set mac
+     * @param mac
+     */
+    withMac(mac: string): Complete;
+}