+++ src/server/hue-lights-api.ts
@@ -1,4 +1,3 @@
-import {Express, Request, Response} from 'express';
 import {EMPTY, merge} from 'rxjs';
 import {catchError, map} from 'rxjs/operators';
 import {HueBuilder} from '../builder/hue-builder';
@@ -6,10 +5,13 @@ import {HueError} from '../error/hue-error';
 import {ErrorResponse} from '../response/error-response';
 import {isDefined} from '../util/utils';
 import {HueServerCallback} from './hue-server-callback';
+import {HueS} from './lib/hue-s';
+import {HueSRequest} from './lib/hue-s-request';
+import {HueSResponse} from './lib/hue-s-response';
 
 export class HueLightsApi {
 
-    constructor(private app: Express, private builder: HueBuilder, private callbacks: HueServerCallback) {
+    constructor(private app: HueS, private builder: HueBuilder, private callbacks: HueServerCallback) {
         // 1. Lights API
         if (this.callbacks.onLights) {
             this.app.get('/api/:username/lights', this.onLights);
@@ -34,31 +36,28 @@ export class HueLightsApi {
         }
     }
 
-    private onLights = (req: Request, res: Response) => {
+    private onLights = (req: HueSRequest, res: HueSResponse) => {
         const username = req.params.username;
-        this.builder.logger.debug(`HueServer: Incoming GET /api/${username}/lights request`);
 
         this.callbacks.onLights!(req, username).subscribe(lights => {
             res.json(lights);
         }, (err: HueError) => {
-            res.json(ErrorResponse.create(err, '/lights'));
+            res.json([ErrorResponse.create(err, '/lights')]);
         });
     };
 
-    private onLightsNew = (req: Request, res: Response) => {
+    private onLightsNew = (req: HueSRequest, res: HueSResponse) => {
         const username = req.params.username;
-        this.builder.logger.debug(`HueServer: Incoming GET /api/${username}/lights/new request`);
 
         this.callbacks.onLightsNew!(req, username).subscribe(lights => {
             res.json(lights);
         }, (err: HueError) => {
-            res.json(ErrorResponse.create(err, '/lights/new'));
+            res.json([ErrorResponse.create(err, '/lights/new')]);
         });
     };
 
-    private onLightsSearchNew = (req: Request, res: Response) => {
+    private onLightsSearchNew = (req: HueSRequest, res: HueSResponse) => {
         const username = req.params.username;
-        this.builder.logger.debug(`HueServer: Incoming POST /api/${username}/lights request:\n${JSON.stringify(req.body)}\n`);
 
         this.callbacks.onLightsSearchNew!(req, username, req.body?.deviceid).subscribe(() => {
             res.json([{
@@ -69,22 +68,20 @@ export class HueLightsApi {
         });
     };
 
-    private onLight = (req: Request, res: Response) => {
+    private onLight = (req: HueSRequest, res: HueSResponse) => {
         const username = req.params.username;
         const lightId = req.params.id;
-        this.builder.logger.debug(`HueServer: Incoming GET /api/${username}/lights/${lightId} request`);
 
         this.callbacks.onLight!(req, username, lightId).subscribe(light => {
             res.json(light);
         }, (err: HueError) => {
-            res.json(ErrorResponse.create(err, `/lights/${lightId}`));
+            res.json([ErrorResponse.create(err, `/lights/${lightId}`)]);
         });
     };
 
-    private onLightsRename = (req: Request, res: Response) => {
+    private onLightsRename = (req: HueSRequest, res: HueSResponse) => {
         const username = req.params.username;
         const lightId = req.params.id;
-        this.builder.logger.debug(`HueServer: Incoming PUT /api/${username}/lights/${lightId} request:\n${JSON.stringify(req.body)}\n`);
 
         if (!req.body?.name) {
             res.json(ErrorResponse.create(HueError.PARAMETER_NOT_AVAILABLE.withParams('name'), `/lights/${lightId}`));
@@ -94,14 +91,13 @@ export class HueLightsApi {
         this.callbacks.onLightsRename!(req, username, lightId, req.body.name).subscribe(light => {
             res.json(light);
         }, (err: HueError) => {
-            res.json(ErrorResponse.create(err, `/lights/${lightId}`));
+            res.json([ErrorResponse.create(err, `/lights/${lightId}`)]);
         });
     };
 
-    private onLightsState = (req: Request, res: Response) => {
+    private onLightsState = (req: HueSRequest, res: HueSResponse) => {
         const username = req.params.username;
         const lightId = req.params.id;
-        this.builder.logger.debug(`HueServer: Incoming PUT /api/${username}/lights/${lightId}/state request:\n${JSON.stringify(req.body)}\n`);
 
         const response: any[] = [];
 
@@ -148,15 +144,14 @@ export class HueLightsApi {
         });
     };
 
-    private onLightsDelete = (req: Request, res: Response) => {
+    private onLightsDelete = (req: HueSRequest, res: HueSResponse) => {
         const username = req.params.username;
         const lightId = req.params.id;
-        this.builder.logger.debug(`HueServer: Incoming DELETE /api/${username}/lights/${lightId} request`);
 
         this.callbacks.onLightsDelete!(req, username, lightId).subscribe(light => {
             res.json(light);
         }, (err: HueError) => {
-            res.json(ErrorResponse.create(err, `/lights/${lightId}`));
+            res.json([ErrorResponse.create(err, `/lights/${lightId}`)]);
         });
     };
 }