+++ src/server/hue-server.ts
@@ -1,8 +1,3 @@
-import * as bodyParser from 'body-parser';
-import express from 'express';
-import {Express, Request, Response} from 'express';
-import * as http from 'http';
-import * as https from 'https';
 import {HueBuilder} from '../builder/hue-builder';
 import {HueError} from '../error/hue-error';
 import {isUndefined} from '../util/utils';
@@ -12,6 +7,10 @@ import {HueLightsApi} from './hue-lights-api';
 import {HueServerCallback} from './hue-server-callback';
 import {ErrorResponse} from '../response/error-response';
 import {discovery} from '../util/discovery';
+import {HueS} from './lib/hue-s';
+import {HueSFastify} from './lib/hue-s-fastify';
+import {HueSRequest} from './lib/hue-s-request';
+import {HueSResponse} from './lib/hue-s-response';
 
 /**
  * Responsible for handling actual calls
@@ -21,25 +20,20 @@ import {discovery} from '../util/discovery';
  */
 export class HueServer {
 
-    private readonly app: Express;
+    private readonly app: HueS;
 
     constructor(private builder: HueBuilder, private callbacks: HueServerCallback) {
-
-        this.app = express();
-        this.app.use(bodyParser.json({type: '*/*'}));
-
-        this.app.use(bodyParser.urlencoded({extended: true}));
-        this.app.use((req, res, next) => {
-            res.append('Access-Control-Allow-Origin', ['*']);
-            res.append('Access-Control-Allow-Methods', 'POST, GET, OPTIONS, PUT, DELETE, HEAD');
-            res.append('Access-Control-Allow-Credentials', 'true');
-            res.append('Access-Control-Max-Age', '3600');
-            res.append('Access-Control-Allow-Headers', 'Origin, X-Requested-With, Content-Type, Accept');
-            res.append('Connection', 'close'); // This is important. Otherwise some clients may fail
-            next();
+        this.app = new HueSFastify(this.builder);
+        this.app.setDefaultResponseHeaders({
+            'Access-Control-Allow-Origin': '*',
+            'Access-Control-Allow-Methods': 'POST, GET, OPTIONS, PUT, DELETE, HEAD',
+            'Access-Control-Allow-Credentials': 'true',
+            'Access-Control-Max-Age': '3600',
+            'Access-Control-Allow-Headers': 'Origin, X-Requested-With, Content-Type, Accept',
+            'Connection': 'close' // This is important. Otherwise some clients may fail
         });
 
-        this.app.get('/api/discovery.xml', this.onDiscovery);
+        this.app.get('/description.xml', this.onDiscovery);
 
         // 1. Lights API
         new HueLightsApi(this.app, this.builder, this.callbacks);
@@ -56,6 +50,8 @@ export class HueServer {
         // 7. Configuration API
         if (this.callbacks.onPairing) {
             this.app.post('/api', this.onPairing);
+            // Harmony does that. But only for pairing. Everything else looks normal.
+            this.app.post('/api/', this.onPairing);
         }
         if (this.callbacks.onConfig) {
             this.app.get('/api/config', this.onConfig);
@@ -73,41 +69,31 @@ export class HueServer {
         // Fallback
         new HueFallback(this.app, this.builder, this.callbacks);
 
-        const httpServer = http.createServer(this.app);
-
-        httpServer.listen(this.builder.port, this.builder.host, () => {
-            this.builder.logger.debug(`HueServer: Http-Server listening ${this.builder.host}:${this.builder.port}`);
+        this.app.startServer(() => {
+            // nothing to do so far
         });
 
-        if (this.builder.httpsConfig) {
-            const options = {
-                key: this.builder.httpsConfig.key,
-                cert: this.builder.httpsConfig.cert,
-                rejectUnauthorized: false,
-            };
-
-            const httpsServer = https.createServer(options, this.app);
-
-            httpsServer.listen(this.builder.httpsConfig.port, this.builder.host, () => {
-                this.builder.logger.debug(`HueServer: Https-Server listening ${this.builder.host}:${this.builder.httpsConfig?.port}`);
-            });
-        }
+        this.app.registerOnRequest(req => {
+            if(req.body) {
+                this.builder.logger.debug(`HueServer: Incoming ${req.method} ${req.url} request from ${req.ip}.\nBody:\n${JSON.stringify(req.body)}`);
+            } else {
+                this.builder.logger.debug(`HueServer: Incoming ${req.method} ${req.url} request from ${req.ip}`);
+            }
+            this.builder.logger.fine(`HueServer: Headers:\n${JSON.stringify(req.headers)}`);
+        });
     }
 
-    private onDiscovery = (req: Request, res: Response) => {
-        this.builder.logger.debug(`HueServer: Incoming discovery request.`);
-        res.contentType('application/xml');
-        res.send(discovery(this.builder.discoveryHost, this.builder.discoveryPort, this.builder.udn));
+    private onDiscovery = (req: HueSRequest, res: HueSResponse) => {
+        res.setContentType('application/xml');
+        res.send(discovery(this.builder.discoveryHost, this.builder.discoveryPort, this.builder.udn, this.builder.mac));
     };
 
-    private onPairing = (req: Request, res: Response) => {
-        this.builder.logger.debug(`HueServer: Incoming pairing request:\n${JSON.stringify(req.body)}\n`);
-
+    private onPairing = (req: HueSRequest, res: HueSResponse) => {
         if (isUndefined(req.body.devicetype)) {
             res.json(ErrorResponse.create(HueError.PARAMETER_NOT_AVAILABLE.withParams('devicetype'), ''));
         }
 
-        this.callbacks.onPairing!(req.body.devicetype, req.body?.generateclientkey).subscribe(username => {
+        this.callbacks.onPairing!(req, req.body.devicetype, req.body?.generateclientkey).subscribe(username => {
             const response = [{
                 success: {
                     username: username
@@ -115,28 +101,24 @@ export class HueServer {
             }];
             res.json(response);
         }, (err: HueError) => {
-            res.json(ErrorResponse.create(err, ''));
+            res.json([ErrorResponse.create(err, '')]);
         });
     };
 
-    private onConfig = (req: Request, res: Response) => {
-        this.builder.logger.debug(`HueServer: Incoming /config request`);
-
+    private onConfig = (req: HueSRequest, res: HueSResponse) => {
         this.callbacks.onConfig!(req).subscribe(config => {
             res.json(config);
         }, (err: HueError) => {
-            res.json(ErrorResponse.create(err, '/config'));
+            res.json([ErrorResponse.create(err, '/config')]);
         });
     };
 
-    private onAll = (req: Request, res: Response) => {
+    private onAll = (req: HueSRequest, res: HueSResponse) => {
         const username = req.params.username;
-        this.builder.logger.debug(`HueServer: Incoming / request by=${username}`);
-
         this.callbacks.onAll!(req, username).subscribe(lights => {
             res.json(lights);
         }, (err: HueError) => {
-            res.json(ErrorResponse.create(err, '/'));
+            res.json([ErrorResponse.create(err, '/')]);
         });
     };
 }