+++ src/server/lib/hue-s-response.ts
@@ -1,0 +1,5 @@
+export interface HueSResponse {
+    setContentType(contentType: string): void;
+    send(data: string): void
+    json(data: object | any[]): void;
+}